# documentation/api-documentation.yaml
openapi: 3.0.0
info:
  title: Fiat to USDC Conversion System API
  description: API documentation for the Fiat to USDC Conversion System
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
  /register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  email:
                    type: string
                  kycStatus:
                    type: string
                  walletAddress:
                    type: string
                  bankVerified:
                    type: boolean
  /transaction:
    post:
      summary: Create a new transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                amount:
                  type: number
                currency:
                  type: string
      responses:
        '201':
          description: Transaction created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  userId:
                    type: string
                  amount:
                    type: number
                  currency:
                    type: string
                  status:
                    type: string
                  exchangeRate:
                    type: number
                  usdcAmount:
                    type: number
  /treasury/operation:
    post:
      summary: Create a new treasury operation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                details:
                  type: object
      responses:
        '201':
          description: Treasury operation created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  type:
                    type: string
                  status:
                    type: string
                  details:
                    type: object
  /monitoring/event:
    post:
      summary: Create a new monitoring event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                details:
                  type: object
      responses:
        '201':
          description: Monitoring event created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  type:
                    type: string
                  status:
                    type: string
                  details:
                    type: object